Created a mechanism for forcing sandbox use
In an effort to ensure that the sandbox environment and not the
production environment is in use, I added a :sandbox boolean to the
configuration. I also added a configuration option to hard-wire the API
URL should that be necessary. Fixed a bug with how YAML config loading
worked, as well.